Switch feature limits to long; add media & feedback
Change domain and application feature/quota types from int to long (DeltaChange, Quota, LimitedFeature.Limit, DTOs and service interfaces) and update limited feature defaults (storage now uses bytes). Add MediaRequest abstraction and AddToGallery command to handle media uploads. Add FeedbackSubmittedEvent (includes rating) and an event handler that updates organization rating with new validation error for out-of-range ratings. Improve UpdateLogo handler to delete previous media and persist changes, and add UnitOfWork.SaveChangesAsync calls to branch handlers. Add EF Core migration to change DB columns to bigint and update the model snapshot. Also adjust OrganizationsController search to accept sortBy and filter active organizations.

+++ Viora/Viora.Application/Abstractions/Messaging/ILimitedFeatureCommand.cs
@@ -11,5 +11,5 @@ public interface IBaseLimitedFeatureCommand
     /// <summary>
     /// represents the amount to be added or removed the quota, +ve means add an amount (restore), negative means consume amount
     /// </summary>
-    public int DeltaChange { get; init; }
+    public long DeltaChange { get; init; }
 }
