+++ Viora/Viora.Domain/Plans/Services/ILimitedFeatureUsageService.cs
@@ -4,7 +4,7 @@ namespace Viora.Domain.Plans.Services;
 
 public interface ILimitedFeatureUsageService
 {
-    Task<Result> CheckLimitAsync(Guid organizationId, Guid limitedFeatureId, int delta, CancellationToken cancellationToken);
-    Task<Result> ConsumeLimit(Guid organizationId, Guid limitedFeatureId, int delta, CancellationToken cancellationToken);
+    Task<Result> CheckLimitAsync(Guid organizationId, Guid limitedFeatureId, long delta, CancellationToken cancellationToken);
+    Task<Result> ConsumeLimit(Guid organizationId, Guid limitedFeatureId, long delta, CancellationToken cancellationToken);
 
 }
